REF: add parent class for accessor generators
